Added TutorialHandler object that scripts some basic sequences for educational purposes.
Made Shipment notify a list of subscribers upon completing a delivery.
Tweaked Shipment to spawn more things faster.
DayValues set to begin with a shipment that contains 5 of each basic item. This is not dynamic and will need to be regenerated if the day1 orders change.
Wiped the starting inventory from PlayerValues.
Remove redundant / unused method from SpawnArea.